+++ src/components/manage/Toolbar/Toolbar.jsx
@@ -7,7 +7,6 @@ import React, { Component } from 'react';
 import { defineMessages, injectIntl } from 'react-intl';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import jwtDecode from 'jwt-decode';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { doesNodeContainClick } from 'semantic-ui-react/dist/commonjs/lib';
@@ -28,6 +27,10 @@ import {
   setExpandedToolbar,
   unlockContent,
 } from '@plone/volto/actions';
+import {
+  loggedIn,
+  userData,
+} from '@plone/volto/selectors/userSession/userSession';
 import { Icon } from '@plone/volto/components';
 import { BodyClass, getBaseUrl } from '@plone/volto/helpers';
 import { Pluggable } from '@plone/volto/components/manage/Pluggable';
@@ -140,8 +143,7 @@ class Toolbar extends Component {
       object_buttons: PropTypes.arrayOf(PropTypes.object),
       user: PropTypes.arrayOf(PropTypes.object),
     }),
-    token: PropTypes.string,
-    userId: PropTypes.string,
+    userLoggedIn: PropTypes.bool,
     pathname: PropTypes.string.isRequired,
     content: PropTypes.shape({
       '@type': PropTypes.string,
@@ -170,8 +172,6 @@ class Toolbar extends Component {
    */
   static defaultProps = {
     actions: null,
-    token: null,
-    userId: null,
     content: null,
     hideDefaultViewButtons: false,
     types: [],
@@ -308,7 +308,7 @@ class Toolbar extends Component {
     const { expanded } = this.state;
 
     return (
-      this.props.token && (
+      this.props.userLoggedIn && (
         <>
           <BodyClass
             className={expanded ? 'has-toolbar' : 'has-toolbar-collapsed'}
@@ -579,10 +579,8 @@ export default compose(
   connect(
     (state, props) => ({
       actions: state.actions.actions,
-      token: state.userSession.token,
-      userId: state.userSession.token
-        ? jwtDecode(state.userSession.token).sub
-        : '',
+      userLoggedIn: loggedIn(state),
+      userId: userData(state).userId,
       content: state.content.data,
       pathname: props.pathname,
       types: filter(state.types.types, 'addable'),