+++ src/components/theme/Anontools/Anontools.jsx
@@ -9,7 +9,9 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { Menu } from 'semantic-ui-react';
 import { FormattedMessage } from 'react-intl';
+
 import config from '@plone/volto/registry';
+import { loggedIn } from '@plone/volto/selectors/userSession/userSession';
 
 /**
  * Anontools container class.
@@ -23,7 +25,7 @@ export class Anontools extends Component {
    * @static
    */
   static propTypes = {
-    token: PropTypes.string,
+    userLoggedIn: PropTypes.bool,
     content: PropTypes.shape({
       '@id': PropTypes.string,
     }),
@@ -35,7 +37,6 @@ export class Anontools extends Component {
    * @static
    */
   static defaultProps = {
-    token: null,
     content: {
       '@id': null,
     },
@@ -49,7 +50,7 @@ export class Anontools extends Component {
   render() {
     const { settings } = config;
     return (
-      !this.props.token && (
+      !this.props.userLoggedIn && (
         <Menu pointing secondary floated="right">
           <Menu.Item>
             <Link
@@ -80,6 +81,6 @@ export class Anontools extends Component {
 }
 
 export default connect((state) => ({
-  token: state.userSession.token,
+  userLoggedIn: loggedIn(state),
   content: state.content.data,
 }))(Anontools);