+++ src/components/theme/Anontools/Anontools.test.jsx
@@ -5,13 +5,17 @@ import { Provider } from 'react-intl-redux';
 import { MemoryRouter } from 'react-router-dom';
 
 import Anontools from './Anontools';
+import { arrayWIdsToObject } from '@plone/volto/helpers/Utils/Utils';
 
 const mockStore = configureStore();
 
+const actions = { user: [{ id: 'login' }] };
+const actionsById = arrayWIdsToObject(actions);
+
 describe('Anontools', () => {
-  it('renders an anontools component when no token is specified', () => {
+  it('renders an anontools component when no use is logged in', () => {
     const store = mockStore({
-      userSession: { token: null },
+      actions: { actions, actionsById },
       content: { data: { '@id': 'myid' } },
       intl: {
         locale: 'en',
@@ -29,9 +33,10 @@ describe('Anontools', () => {
     expect(json).toMatchSnapshot();
   });
 
-  it('should not render an anontools component when a token is specified', () => {
+  it('should not render an anontools component when a user is logged in', () => {
+    const actions = { user: [{ id: 'logout' }] };
     const store = mockStore({
-      userSession: { token: '1234' },
+      actions: { actions, actionsById: arrayWIdsToObject(actions) },
       content: { data: {} },
       intl: {
         locale: 'en',