+++ src/components/theme/View/View.test.jsx
@@ -5,6 +5,7 @@ import { Provider } from 'react-intl-redux';
 
 import View from './View';
 import config from '@plone/volto/registry';
+import { arrayWIdsToObject } from '@plone/volto/helpers/Utils/Utils';
 
 beforeAll(() => {
   config.set('views', {
@@ -129,18 +130,18 @@ const actions = {
     },
     {
       icon: '',
-      id: 'logout',
-      title: 'Log out',
+      id: 'login',
+      title: 'Log in',
     },
   ],
 };
+const actionsById = arrayWIdsToObject(actions);
 
 describe('View', () => {
   it('renders an empty view', () => {
     const store = mockStore({
-      actions: { actions },
+      actions: { actions, actionsById },
       content: { get: { error: null } },
-      userSession: { token: null },
       apierror: {},
       intl: {
         locale: 'en',
@@ -158,9 +159,8 @@ describe('View', () => {
 
   it('renders a summary view', () => {
     const store = mockStore({
-      actions: { actions },
+      actions: { actions, actionsById },
       content: { data: { layout: 'summary_view' }, get: { error: null } },
-      userSession: { token: null },
       apierror: {},
       intl: {
         locale: 'en',
@@ -178,9 +178,8 @@ describe('View', () => {
 
   it('renders a tabular view', () => {
     const store = mockStore({
-      actions: { actions },
+      actions: { actions, actionsById },
       content: { data: { layout: 'tabular_view' }, get: { error: null } },
-      userSession: { token: null },
       apierror: {},
       intl: {
         locale: 'en',
@@ -198,9 +197,8 @@ describe('View', () => {
 
   it('renders a document view', () => {
     const store = mockStore({
-      actions: { actions },
+      actions: { actions, actionsById },
       content: { data: {}, get: { error: null } },
-      userSession: { token: null },
       apierror: {},
       intl: {
         locale: 'en',